Moved folders to fix git casing
